Prøver oppdatering av Eures data
